Change main body to a template
Instead of setting visibility to hidden the rest
of the elements at first, setting them to a template
works better because it doesn't take up container space
unlike the first approach where even though only the header
and the play button are visible, the container itself is way
too big compared to what's needed. With the template approach,
the container size is appropriate to the visible content.